# api-spec.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/upload/presigned-url":  {
                                                       "post":  {
                                                                    "operationId":  "UploadController_generatePresignedUrl",
                                                                    "parameters":  [

                                                                                   ],
                                                                    "requestBody":  {
                                                                                        "required":  true,
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/GeneratePresignedUrlDto"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  ""
                                                                                              }
                                                                                  }
                                                                }
                                                   },
...
                  "/api/v1/upload/file-info":  {
                                                   "get":  {
                                                               "operationId":  "UploadController_getFileInfo",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "fileUrl",
                                                                                      "required":  true,
                                                                                      "in":  "query",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  ""
                                                                                         }
                                                                             }
                                                           }
                                               },
...
              },
    "info":  {
                 "title":  "Castlyo API",
                 "description":  "Professional casting and talent matching platform API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "GeneratePresignedUrlDto":  {
                                                                       "type":  "object",
                                                                       "properties":  {

                                                                                      }
                                                                   },
...
                                   }
                   }
}